+++ Runtime/Scripts/System/Extensions/Numerics/FloatExtensions.cs
@@ -158,9 +158,15 @@
 			return value.Round() + Float.OneHalf;
 		}
 
-		public static int Sign(this float value)
+		/// <summary>
+		/// Returns <c>-1f</c>, <c>0f</c>, or <c>1f</c> to indicate the <c>sign</c> of the <c>value</c>.
+		/// </summary>
+		/// <remarks>
+		/// See https://docs.microsoft.com/dotnet/api/system.math.sign
+		/// </remarks>
+		public static float Sign(this float value)
 		{
-			return Math.Sign(value);
+			return (float)Math.Sign(value);
 		}
 
 		public static bool ToBool(this float value)
