+++ Runtime/Scripts/System/Extensions/Numerics/ShortExtensions.cs
@@ -62,9 +62,15 @@ namespace WellDefinedValues
 				remainder + divisor : remainder);
 		}
 
-		public static int Sign(this short value)
+		/// <summary>
+		/// Returns <c>-1</c>, <c>0</c>, or <c>1</c> to indicate the <c>sign</c> of the <c>value</c>.
+		/// </summary>
+		/// <remarks>
+		/// See https://docs.microsoft.com/dotnet/api/system.math.sign
+		/// </remarks>
+		public static short Sign(this short value)
 		{
-			return Math.Sign(value);
+			return (short)Math.Sign(value);
 		}
 
 		public static string ToBinaryString(this short value)
